Add button to clear list and move component to seperate file
