+++ src/component/App.js
@@ -1,0 +1,46 @@
+import { useState } from "react";
+import Logo from "./Logo";
+import Form from "./Form";
+import PackingList from "./PackingList";
+import Item from "./Item";
+import Stats from "./Stats";
+
+export default function App() {
+  const [item, setItem] = useState([]);
+
+  function handleAddItem(item) {
+    setItem((items) => [...items, item]);
+  }
+
+  function handleDeleteItem(id) {
+    console.log(id);
+    setItem((items) => items.filter((i) => i.id !== id));
+  }
+
+  function handleToggleItem(id) {
+    setItem((items) =>
+      items.map((item) =>
+        item.id === id ? { ...item, packed: !item.packed } : item
+      )
+    );
+  }
+
+  function handleClearItems() {
+    const confirmed = window.confirm("Are you sure you want clear list items?");
+    if (confirmed) setItem([]);
+  }
+
+  return (
+    <div className="app">
+      <Logo />
+      <Form onAddItem={handleAddItem} />
+      <PackingList
+        items={item}
+        onDeleteItem={handleDeleteItem}
+        onToggleItem={handleToggleItem}
+        onClearItem={handleClearItems}
+      />
+      <Stats item={item} />
+    </div>
+  );
+}
