+++ src/component/Form.js
@@ -1,0 +1,45 @@
+import { useState } from "react";
+
+export default function Form({ onAddItem }) {
+  const [description, setDescription] = useState("");
+  const [quantity, setQuantity] = useState(1);
+
+  return (
+    <form className="add-form" onSubmit={handleSubmit}>
+      <h3>What do you need for your trip?</h3>
+      <select
+        value={quantity}
+        onChange={(e) => setQuantity(Number(e.target.value))}
+      >
+        {Array.from({ length: 20 }, (_, i) => i + 1).map((num) => (
+          <option value={num} key={num}>
+            {num}
+          </option>
+        ))}
+      </select>
+      <input
+        type="text"
+        placeholder={" item..."}
+        value={description}
+        onChange={(e) => setDescription(e.target.value)}
+      />
+      <button>Add</button>
+    </form>
+  );
+
+  function handleSubmit(e) {
+    e.preventDefault();
+    const newItem = {
+      description: description,
+      quantity: quantity,
+      packed: false,
+      id: Date.now(),
+    };
+    console.log(newItem);
+
+    onAddItem(newItem);
+
+    setDescription("");
+    setQuantity(1);
+  }
+}
