+++ src/component/Logo.js
@@ -1,0 +1,7 @@
+export default function Logo() {
+  return (
+    <header>
+      <h1>🏝️ Far Away 🧳</h1>
+    </header>
+  );
+}
