+++ src/index.css
@@ -7,7 +7,7 @@
 }
 
 html {
-  font-size: 62.5%;
+  font-size: 50%;
 }
 
 body {
@@ -29,7 +29,7 @@ h1 {
   text-align: center;
   background-color: #f4a226;
   font-family: "Monoton";
-  font-size: 8rem;
+  font-size: 6rem;
   text-transform: uppercase;
   font-weight: 400;
   word-spacing: 30px;
